Brings in all work from jetson-optimization-A branch (full history preserved):

Key outcomes:
- YOLO on GPU via TensorRT FP16 engine (built natively on Jetson, SM 8.7)
- VLM on CPU async worker (NvMap memory constraint — see JETSON_OPTIMIZATION.md)
- Pipeline: 7.94 fps (PyTorch .pt) → 22.24 fps (TRT FP16 + single YOLO/frame)
- YOLO capacity: 59 ms → 28 ms/frame (2.1× via TRT), now exceeds 30 fps source

Notable commits included:
- feat: unblock GPU inference, NvMap workaround, first benchmark (7.94 fps)
- feat: TensorRT FP16 engine build — 59ms→28ms YOLO, 7.9→15.0 fps pipeline
- feat: INT8 investigation — FP16 640x640 confirmed optimal for SM 8.7
- perf: eliminate double YOLO per frame — 14.95→22.24 fps
- docs: JETSON_OPTIMIZATION.md rewrite + pipeline README Jetson section
- chore: repo cleanup (remove session scratchpads, NGC artifacts)

New config keys: config_vlm_device, config_yolo_imgsz
New files: config.jetson.yaml, docker/setup-native-jetson.sh, scripts/build_trt_int8.py

+++ src/vlm-layer/layer.py
@@ -124,7 +124,10 @@ def initialize_vlm_layer(config: VLMConfig) -> VLMRuntimeState:
     _maybe_require_newer_transformers_for_checkpoint(model_path=model_path)
 
     runtime_device = _resolve_device(torch=torch, requested_device=config.config_device)
-    runtime_dtype = torch.float16 if runtime_device == "cuda" else torch.float32
+    # Use bfloat16 on CUDA — this is the model's native dtype and avoids a
+    # float16 conversion during weight loading that can trigger the Jetson
+    # NvMap / CUDACachingAllocator bug on unified-memory devices.
+    runtime_dtype = torch.bfloat16 if runtime_device == "cuda" else torch.float32
 
     try:
         processor = AutoProcessor.from_pretrained(
